refactor index.ts and create server.ts
